requirements.txt
asttokens==2.0.4
boto3==1.17.18
botocore==1.20.18
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
executing==0.5.4
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
google-auth==1.27.0
google-auth-oauthlib==0.4.2
gspread==3.7.0
gunicorn==20.0.4
icecream==2.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
MarkupSafe==1.1.1
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.8.0
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.3.4
six==1.15.0
SQLAlchemy==1.3.23
urllib3==1.26.3
Werkzeug==1.0.1
WTForms==2.3.3
